Update monitorstyle.css
added the .dark-theme and .light-theme classes to make the dark/light theme button work. if any changes to font or anything else needs to be made in the other pages please include in here. the .dark-theme is commented out on purpose, it's not really needed because all the dark mode features are already coded in the default settings. this code can also be copy and paste.